requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.11
Django==3.1.6
django-extensions==3.1.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
idna==2.10
mysqlclient==2.0.3
nexmo==2.5.2
phonenumbers==8.12.17
pkg-resources==0.0.0
PyJWT==1.7.1
pyotp==2.5.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.51.1
urllib3==1.26.3
wrapt==1.12.1
